examples/google-adk/requirements.txt
google-adk>=1.2.0
